got rid of dashboard errors and cloudinary to simplify createPost process for development
