+++ client/src/pages/Dashboard.js
@@ -1,12 +1,10 @@
 import Feed from '../components/Feed';
-import NewPost from '../components/NewPost';
+import PostForm from '../components/PostForm';
 
 export default function Dashboard() {
   return (
     <div>
-      <button type="button">
-        <NewPost />
-      </button>
+      <PostForm />
       <h2>Dashboard</h2>
       <Feed />
     </div>
